+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### New features
 - [PR #1619](https://github.com/rqlite/rqlite/pull/1619), [PR #1622](https://github.com/rqlite/rqlite/pull/1622): Support automatic `VACUUM` of the SQLite database. Fixes [#1609](https://github.com/rqlite/rqlite/issues/1609).
 
+### Implementation changes and bug fixes
+- [PR #1623](https://github.com/rqlite/rqlite/pull/1623): Unit tests for Swappable DB.
+
 ## 8.16.8 (January 20th 2024)
 ### Implementation changes and bug fixes
 - [PR #1615](https://github.com/rqlite/rqlite/pull/1615): Add extensive WAL checkpoint test at the database level.